[update] users配下のディレクトリ一括変更
